requirements.txt
# FastAPI and web server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
supabase==2.3.4
httpx>=0.24.0,<0.26.0

# AI and ML
langchain==0.1.9
langchain-community==0.0.24
groq==0.4.2
ollama==0.1.6
TTS==0.22.0
PyMuPDF==1.23.26
numpy==1.26.4

# Audio processing
sounddevice==0.4.6
soundfile==0.12.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
